#127: 환전 요청 바디 수정

+++ src/main/java/com/ssafy/ssashinsa/heyfy/exchange/dto/exchange/AccountBalanceResponseDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AccountBalanceResponseDto {
     private String  accountNo;
-    private int accountBalance;
+    private Double accountBalance;
     private String currency;
     private boolean isForeign;
 }
